refactor: split AccountScrape behind facade
Phase 12e drains the 217-eff-line AccountScrapeStrategy god-module
into co-located <=150-line siblings behind an unchanged facade:

- AccountScrapeShared: URL/field leaf helpers (patchUrlRange,
  txnEpForParse).
- AccountScrapePost: POST range/direct fetch + context builder
  (scrapePostWithRange, scrapePostDirect, buildPostCtx).
- AccountScrapeFirstWave: DASHBOARD-harvest fast path (tryFirstWave).

AccountScrapeStrategy keeps only the two public orchestrators
(scrapeOneAccountPost, scrapeOneAccountViaUrl) and re-imports from the
siblings. Public surface and behavior are byte-identical; all 94
Account unit tests pass unmodified.
